+++ lib/active_record_query_counter/sidekiq_middleware.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module ActiveRecordQueryCounter
+  # Sidekiq middleware to count queries on a job.
+  class SidekiqMiddleware
+    def call(worker, job, queue, &block)
+      ActiveRecordQueryCounter.count_queries(&block)
+    end
+  end
+end