+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.0.0
 
 ### Changed
 
 - Calculate elapsed time using monotonic time rather than wall clock time.
+- Added method to get the amount of time a single transaction could have taken if it was used to wrap multiple updates.
+- Breaking change: transaction information is now returned in a `ActiveRecordQueryCounter::TransactionInfo` objects rather than as a hash of arrays.
 
 ## 1.1.2